Attempt 3: Bin particles then parallelising
Each box size is 5 * radius.

Each box check collision with left
neighbour, then right neighbour ... and so on.

In iteration of checking a bin collisions with its neighbour,
this can be parallelise, so other threads can do the same for other bins.

By choosing a good bin size, can be proven to have no race codition.